move prefixes to mongo, remove enable/disable
